tweak example for OpenCV
